+++ src/main/java/src/model/BuildingsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BuildingsRepository extends JpaRepository<Building, Long> {
+public interface BuildingsRepository extends JpaRepository<Building, Integer> {
 
     List<Building> findByName(String name);
